✨ Add welcome Page as loading screen Forwards after three seconds
